src/python_application/requirements.txt
requests == 2.26.0
evolved5g
configparser
redis
pyopenssl
pyjwt
termcolor